fixed music
